requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.4
# pandas==1.3.4
requests==2.26.0
Pillow==8.4.0
psycopg2-binary==2.9.2
Django==3.2.9
django-compat==1.0.15
django-filter==21.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4


